Split song format parser into Song.Format module.
